@@ -2,6 +2,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.IO;
 using System.Linq;
 using FluentAssertions;
+using ShapeCrawler.Tests.Shared;
 using ShapeCrawler.UnitTests.Helpers;
 using ShapeCrawler.UnitTests.Properties;
 using Xunit;
@@ -19,7 +20,7 @@ public class SlideTests : ShapeCrawlerTest
     public void Hide_MethodHidesSlide_WhenItIsExecuted()
     {
         // Arrange
-        var pptx = GetTestStream("001.pptx");
+        var pptx = Assets.GetStream("001.pptx");
         var pre = SCPresentation.Open(pptx);
         var slide = pre.Slides.First();
 
@@ -67,7 +68,6 @@ var pre = SCPresentation.Open(pptx);
     public void Background_ImageIsNull_WhenTheSlideHasNotBackground()
     {
         // Arrange
-        var pres9 = SCPresentation.Open(GetTestStream("009_table.pptx"));
         var slide = SCPresentation.Open(GetTestStream("009_table.pptx")).Slides[1];
 
         // Act
@@ -82,7 +82,7 @@ var pre = SCPresentation.Open(pptx);
     {
         // Arrange
         const string customDataString = "Test custom data";
-        var originPre = SCPresentation.Open(GetTestStream("001.pptx"));
+        var originPre = SCPresentation.Open(Assets.GetStream("001.pptx"));
         var slide = originPre.Slides.First();
 
         // Act
@@ -101,8 +101,7 @@ var pre = SCPresentation.Open(pptx);
     public void CustomData_PropertyIsNull_WhenTheSlideHasNotCustomData()
     {
         // Arrange
-        var pres1 = SCPresentation.Open(GetTestStream("001.pptx"));
-        var slide = SCPresentation.Open(GetTestStream("001.pptx")).Slides.First();
+        var slide = SCPresentation.Open(Assets.GetStream("001.pptx")).Slides.First();
 
         // Act
         var sldCustomData = slide.CustomData;
@@ -115,7 +114,7 @@ var pre = SCPresentation.Open(pptx);
     public void Number_Setter_moves_slide_to_specified_number_position()
     {
         // Arrange
-        var pptxStream = TestFiles.Presentations.pre001_stream;
+        var pptxStream = Assets.GetStream("001.pptx");
         var pres = SCPresentation.Open(pptxStream);
         var slide1 = pres.Slides[0];
         var slide2 = pres.Slides[1];
