@@ -5,6 +5,7 @@ using System.IO;
 using System.Linq;
 using FluentAssertions;
 using ShapeCrawler.Shapes;
+using ShapeCrawler.Tests.Shared;
 using ShapeCrawler.UnitTests.Helpers;
 using ShapeCrawler.UnitTests.Helpers.Attributes;
 using ShapeCrawler.UnitTests.Helpers;
@@ -85,7 +86,7 @@ public class TableTests : ShapeCrawlerTest
     public void Row_Height_Getter_returns_height_of_row()
     {
         // Arrange
-        var pptx = GetTestStream("001.pptx");
+        var pptx = Assets.GetStream("001.pptx");
         var pres = SCPresentation.Open(pptx);
         var table = (ITable)pres.Slides[1].Shapes.First(sp => sp.Id == 3);
         
@@ -108,17 +109,15 @@ public class TableTests : ShapeCrawlerTest
 
     public static IEnumerable<object[]> TestCasesCellIsMergedCell()
     {
-        var pptx1 = GetTestStream("001.pptx");
-        var pres1 = SCPresentation.Open(pptx1);
-        var table1 = SCPresentation.Open(GetTestStream("001.pptx")).Slides[1].Shapes.GetById<ITable>(3);
+        var table1 = SCPresentation.Open(Assets.GetStream("001.pptx")).Slides[1].Shapes.GetById<ITable>(3);
         yield return new object[] {table1[0, 0], table1[1, 0]};
         
-        var pptx2 = GetTestStream("001.pptx");
+        var pptx2 = Assets.GetStream("001.pptx");
         var pres2 = SCPresentation.Open(pptx2);
         var table2 = pres2.Slides[1].Shapes.GetByName<ITable>("Table 5");
         yield return new object[] {table2[1, 1], table2[2, 1]};
         
-        var pptx3 = GetTestStream("001.pptx");
+        var pptx3 = Assets.GetStream("001.pptx");
         var pres3 = SCPresentation.Open(pptx3);
         var table3 = pres3.Slides[3].Shapes.GetById<ITable>(4);
         yield return new object[] {table3[0, 1], table3[1, 1]};
@@ -128,8 +127,7 @@ public class TableTests : ShapeCrawlerTest
     public void ColumnsCount_ReturnsNumberOfColumnsInTheTable()
     {
         // Arrange
-        var pres1 = SCPresentation.Open(GetTestStream("001.pptx"));
-        ITable table = (ITable)SCPresentation.Open(GetTestStream("001.pptx")).Slides[1].Shapes.First(sp => sp.Id == 4);
+        ITable table = (ITable)SCPresentation.Open(Assets.GetStream("001.pptx")).Slides[1].Shapes.First(sp => sp.Id == 4);
 
         // Act
         int columnsCount = table.Columns.Count;
@@ -142,8 +140,7 @@ public class TableTests : ShapeCrawlerTest
     public void Column_Width_Getter_returns_width_of_column_in_pixels()
     {
         // Arrange
-        var pres1 = SCPresentation.Open(GetTestStream("001.pptx"));
-        var table = (ITable)SCPresentation.Open(GetTestStream("001.pptx")).Slides[1].Shapes.First(sp => sp.Id == 4);
+        var table = (ITable)SCPresentation.Open(Assets.GetStream("001.pptx")).Slides[1].Shapes.First(sp => sp.Id == 4);
 
         // Act
         var columnWidth = table.Columns[0].Width;
@@ -156,7 +153,7 @@ public class TableTests : ShapeCrawlerTest
     public void Column_Width_Setter_sets_width_of_column()
     {
         // Arrange
-        var pres = SCPresentation.Open(GetTestStream("001.pptx"));
+        var pres = SCPresentation.Open(Assets.GetStream("001.pptx"));
         var table = (ITable)pres.Slides[1].Shapes.First(sp => sp.Id == 3);
         const int newColumnWidth = 427;
         var mStream = new MemoryStream();
@@ -177,7 +174,7 @@ public class TableTests : ShapeCrawlerTest
     public void Row_Cell_IsMergedCell_returns_True_When_cell_is_merged()
     {
         // Arrange
-        var pptx = GetTestStream("001.pptx");
+        var pptx = Assets.GetStream("001.pptx");
         var pres = SCPresentation.Open(pptx);
         var row = pres.Slides[1].Shapes.GetByName<ITable>("Table 4").Rows[1];
         var cell1X0 = row.Cells[0];
@@ -219,7 +216,7 @@ public class TableTests : ShapeCrawlerTest
     public void MergeCells_MergesSpecifiedCellsRange(int rowIdx1, int colIdx1, int rowIdx2, int colIdx2)
     {
         // Arrange
-        IPresentation presentation = SCPresentation.Open(GetTestStream("001.pptx"));
+        IPresentation presentation = SCPresentation.Open(Assets.GetStream("001.pptx"));
         ITable table = (ITable)presentation.Slides[1].Shapes.First(sp => sp.Id == 4);
         var mStream = new MemoryStream();
 
@@ -241,7 +238,7 @@ public class TableTests : ShapeCrawlerTest
     public void MergeCells_Merges0x0And0x1CellsOf2x2Table()
     {
         // Arrange
-        IPresentation presentation = SCPresentation.Open(GetTestStream("001.pptx"));
+        IPresentation presentation = SCPresentation.Open(Assets.GetStream("001.pptx"));
         ITable table = (ITable)presentation.Slides[2].Shapes.First(sp => sp.Id == 5);
         var mStream = new MemoryStream();
 
@@ -265,7 +262,7 @@ public class TableTests : ShapeCrawlerTest
     public void MergeCells_Merges0x1And0x2CellsOf3x2Table()
     {
         // Arrange
-        IPresentation presentation = SCPresentation.Open(GetTestStream("001.pptx"));
+        IPresentation presentation = SCPresentation.Open(Assets.GetStream("001.pptx"));
         ITable table = (ITable)presentation.Slides[2].Shapes.First(sp => sp.Id == 3);
         var mStream = new MemoryStream();
 
@@ -291,7 +288,7 @@ public class TableTests : ShapeCrawlerTest
     public void MergeCells_Merges0x0And0x1And0x2CellsOf3x2Table()
     {
         // Arrange
-        IPresentation presentation = SCPresentation.Open(GetTestStream("001.pptx"));
+        IPresentation presentation = SCPresentation.Open(Assets.GetStream("001.pptx"));
         ITable table = (ITable)presentation.Slides[2].Shapes.First(sp => sp.Id == 3);
         var mStream = new MemoryStream();
 
@@ -315,7 +312,7 @@ public class TableTests : ShapeCrawlerTest
     public void MergeCells_Merges0x0And0x1MergedCellsWith0x2CellIn3x2Table()
     {
         // Arrange
-        IPresentation presentation = SCPresentation.Open(GetTestStream("001.pptx"));
+        IPresentation presentation = SCPresentation.Open(Assets.GetStream("001.pptx"));
         ITable table = (ITable)presentation.Slides[2].Shapes.First(sp => sp.Id == 7);
         var mStream = new MemoryStream();
 
@@ -339,7 +336,7 @@ public class TableTests : ShapeCrawlerTest
     public void MergeCells_merges_0x0_and_1x0_cells_of_2x2_table()
     {
         // Arrange
-        var pptx = GetTestStream("001.pptx");
+        var pptx = Assets.GetStream("001.pptx");
         var pres = SCPresentation.Open(pptx);
         var table = pres.Slides[2].Shapes.GetById<ITable>(5);
         var mStream = new MemoryStream();
@@ -368,7 +365,7 @@ public class TableTests : ShapeCrawlerTest
     public void MergeCells_merges_cells()
     {
         // Arrange
-        var pptx = GetTestStream("001.pptx");
+        var pptx = Assets.GetStream("001.pptx");
         var pres = SCPresentation.Open(pptx);
         var table = pres.Slides[2].Shapes.GetById<ITable>(5);
 
@@ -383,7 +380,7 @@ public class TableTests : ShapeCrawlerTest
     public void MergeCells_Merges0x1And1x1CellsOf3x2Table()
     {
         // Arrange
-        IPresentation presentation = SCPresentation.Open(GetTestStream("001.pptx"));
+        IPresentation presentation = SCPresentation.Open(Assets.GetStream("001.pptx"));
         ITable table = (ITable)presentation.Slides[2].Shapes.First(sp => sp.Id == 3);
         var mStream = new MemoryStream();
 
@@ -407,7 +404,7 @@ public class TableTests : ShapeCrawlerTest
     public void MergeCells_Merges0x0To1x1RangeOf3x3Table()
     {
         // Arrange
-        IPresentation presentation = SCPresentation.Open(GetTestStream("001.pptx"));
+        IPresentation presentation = SCPresentation.Open(Assets.GetStream("001.pptx"));
         ITable table = (ITable)presentation.Slides[2].Shapes.First(sp => sp.Id == 10);
         var mStream = new MemoryStream();
 
@@ -435,7 +432,7 @@ public class TableTests : ShapeCrawlerTest
     public void MergeCells_MergesMergedCellWithNonMergedCell()
     {
         // Arrange
-        IPresentation presentation = SCPresentation.Open(GetTestStream("001.pptx"));
+        IPresentation presentation = SCPresentation.Open(Assets.GetStream("001.pptx"));
         ITable table = (ITable)presentation.Slides[1].Shapes.First(sp => sp.Id == 5);
         var mStream = new MemoryStream();
 
@@ -461,7 +458,7 @@ public class TableTests : ShapeCrawlerTest
     public void MergeCells_MergesTwoMergedCells()
     {
         // Arrange
-        IPresentation presentation = SCPresentation.Open(GetTestStream("001.pptx"));
+        IPresentation presentation = SCPresentation.Open(Assets.GetStream("001.pptx"));
         ITable table = (ITable)presentation.Slides[3].Shapes.First(sp => sp.Id == 2);
         var mStream = new MemoryStream();
 
@@ -491,7 +488,7 @@ public class TableTests : ShapeCrawlerTest
     public void MergeCells_converts_2X1_table_into_1X1_when_all_cells_are_merged()
     {
         // Arrange
-        var pres = SCPresentation.Open(GetTestStream("001.pptx"));
+        var pres = SCPresentation.Open(Assets.GetStream("001.pptx"));
         var table = (ITable)pres.Slides[3].Shapes.First(sp => sp.Id == 3);
         var mStream = new MemoryStream();
         var totalColWidth = table.Columns[0].Width + table.Columns[1].Width;
@@ -518,7 +515,7 @@ public class TableTests : ShapeCrawlerTest
     public void MergeCells_converts_2X2_table_into_1X1_when_all_cells_are_merged()
     {
         // Arrange
-        var pres = SCPresentation.Open(GetTestStream("001.pptx"));
+        var pres = SCPresentation.Open(Assets.GetStream("001.pptx"));
         var table = (ITable)pres.Slides[2].Shapes.First(sp => sp.Id == 5) ;
         var mStream = new MemoryStream();
         var mergedColumnWidth = table.Columns[0].Width + table.Columns[1].Width;
@@ -549,7 +546,7 @@ public class TableTests : ShapeCrawlerTest
     public void MergeCells_merges_0x0_And_0x1_cells_in_3x1_table()
     {
         // Arrange
-        var presentation = SCPresentation.Open(GetTestStream("001.pptx"));
+        var presentation = SCPresentation.Open(Assets.GetStream("001.pptx"));
         var table = (ITable)presentation.Slides[3].Shapes.First(sp => sp.Id == 6) ;
         var mStream = new MemoryStream();
         var mergedColumnWidth = table.Columns[0].Width + table.Columns[1].Width;
@@ -578,7 +575,7 @@ public class TableTests : ShapeCrawlerTest
     public void MergeCells_merges_0x1_and_0x2_cells()
     {
         // Arrange
-        var pptx = GetTestStream("001.pptx");
+        var pptx = Assets.GetStream("001.pptx");
         var pres = SCPresentation.Open(pptx);
         var table = pres.Slides[3].Shapes.GetById<ITable>(6);
         var mStream = new MemoryStream();
@@ -607,7 +604,7 @@ public class TableTests : ShapeCrawlerTest
     public void MergeCells_updates_columns_count()
     {
         // Arrange
-        var pptx = GetTestStream("001.pptx");
+        var pptx = Assets.GetStream("001.pptx");
         var pres = SCPresentation.Open(pptx);
         var table = pres.Slides[3].Shapes.GetById<ITable>(6);
 
@@ -624,8 +621,8 @@ public class TableTests : ShapeCrawlerTest
     public void Indexer_ReturnsCellByRowAndColumnIndexes()
     {
         // Arrange
-        ITable tableCase1 = (ITable)SCPresentation.Open(GetTestStream("001.pptx")).Slides[1].Shapes.First(sp => sp.Id == 4);
-        ITable tableCase2 = (ITable)SCPresentation.Open(GetTestStream("001.pptx")).Slides[3].Shapes.First(sp => sp.Id == 4);
+        ITable tableCase1 = (ITable)SCPresentation.Open(Assets.GetStream("001.pptx")).Slides[1].Shapes.First(sp => sp.Id == 4);
+        ITable tableCase2 = (ITable)SCPresentation.Open(Assets.GetStream("001.pptx")).Slides[3].Shapes.First(sp => sp.Id == 4);
 
         // Act
         ICell scCellCase1 = tableCase1[0, 0];
