@@ -6,6 +6,7 @@ using System.IO;
 using System.Linq;
 using FluentAssertions;
 using ShapeCrawler.Shapes;
+using ShapeCrawler.Tests.Shared;
 using ShapeCrawler.UnitTests.Helpers;
 using ShapeCrawler.UnitTests.Helpers.Attributes;
 using ShapeCrawler.UnitTests.Helpers;
@@ -25,12 +26,12 @@ namespace ShapeCrawler.UnitTests
             // Arrange
             var pptx8 = GetTestStream("008.pptx");
             var pres8 = SCPresentation.Open(pptx8);
-            var pptx1 = GetTestStream("001.pptx");
+            var pptx1 = Assets.GetStream("001.pptx");
             var pres1 = SCPresentation.Open(pptx1);
             var pptx9 = GetTestStream("009_table.pptx");
             var pres9 = SCPresentation.Open(pptx9);
             var textFrame1 = ((IAutoShape)SCPresentation.Open(GetTestStream("008.pptx")).Slides[0].Shapes.First(sp => sp.Id == 3)).TextFrame;
-            var textFrame2 = ((ITable)SCPresentation.Open(GetTestStream("001.pptx")).Slides[1].Shapes.First(sp => sp.Id == 3)).Rows[0].Cells[0]
+            var textFrame2 = ((ITable)SCPresentation.Open(Assets.GetStream("001.pptx")).Slides[1].Shapes.First(sp => sp.Id == 3)).Rows[0].Cells[0]
                 .TextFrame;
             var textFrame3 = ((ITable)SCPresentation.Open(GetTestStream("009_table.pptx")).Slides[2].Shapes.First(sp => sp.Id == 3)).Rows[0].Cells[0]
                 .TextFrame;
@@ -165,7 +166,7 @@ namespace ShapeCrawler.UnitTests
                 
                 var case1 = new TestElementQuery
                 {
-                    Presentation = SCPresentation.Open(GetTestStream("001.pptx")),
+                    Presentation = SCPresentation.Open(Assets.GetStream("001.pptx")),
                     SlideIndex = 0,
                     ShapeId = 3
                 };
@@ -181,7 +182,7 @@ namespace ShapeCrawler.UnitTests
                 
                 var case3 = new TestElementQuery
                 {
-                    Presentation = SCPresentation.Open(GetTestStream("001.pptx")),
+                    Presentation = SCPresentation.Open(Assets.GetStream("001.pptx")),
                     SlideNumber = 2,
                     ShapeName = "Header 1",
                 };
@@ -211,7 +212,7 @@ namespace ShapeCrawler.UnitTests
         public void Text_Setter_updates_text_box_content_and_Reduces_font_size_When_text_is_Overflow()
         {
             // Arrange
-            var pptxStream = GetTestStream("001.pptx");
+            var pptxStream = Assets.GetStream("001.pptx");
             var pres = SCPresentation.Open(pptxStream);
             var textBox = pres.Slides[0].Shapes.GetByName<IAutoShape>("TextBox 8");
             var textFrame = textBox.TextFrame;
@@ -368,7 +369,7 @@ namespace ShapeCrawler.UnitTests
         public void AutofitType_Getter_returns_text_autofit_type()
         {
             // Arrange
-            var pptx = GetTestStream("001.pptx");
+            var pptx = Assets.GetStream("001.pptx");
             var pres = SCPresentation.Open(pptx);
             var autoShape = pres.Slides[0].Shapes.GetById<IAutoShape>(9);
             var textBox = autoShape.TextFrame;
@@ -442,7 +443,7 @@ namespace ShapeCrawler.UnitTests
                 var testCase3 = new TestCase<ITextFrame, int>(3, textBox3, 2);
                 yield return new object[] { testCase3 };
 
-                var pptxStream4 = GetTestStream("001.pptx");
+                var pptxStream4 = Assets.GetStream("001.pptx");
                 var pres4 = SCPresentation.Open(pptxStream4);
                 var autoShape4 = pres4.Slides[1].Shapes.GetById<IAutoShape>(2);
                 var textBox4 = autoShape4.TextFrame;
@@ -457,7 +458,7 @@ namespace ShapeCrawler.UnitTests
             // Arrange
             const string TEST_TEXT = "ParagraphsAdd";
             var mStream = new MemoryStream();
-            var pres = SCPresentation.Open(GetTestStream("001.pptx"));
+            var pres = SCPresentation.Open(Assets.GetStream("001.pptx"));
             var textFrame = ((IAutoShape)pres.Slides[0].Shapes.First(sp => sp.Id == 4)).TextFrame;
             int originParagraphsCount = textFrame.Paragraphs.Count;
 
@@ -495,7 +496,7 @@ namespace ShapeCrawler.UnitTests
         [Fact]
         public void Paragraphs_Add_adds_new_text_paragraph_at_the_end_And_returns_added_paragraph_When_it_has_been_added_after_text_frame_changed()
         {
-            var pres = SCPresentation.Open(GetTestStream("001.pptx"));
+            var pres = SCPresentation.Open(Assets.GetStream("001.pptx"));
             var autoShape = (IAutoShape)pres.Slides[0].Shapes.First(sp => sp.Id == 3);
             var textBox = autoShape.TextFrame;
             var paragraphs = textBox.Paragraphs;
