@@ -11,9 +11,9 @@ public interface IRoundedRectangle : IAutoShape
 {
 }
 
-internal sealed class SCRoundedRectangle : AutoSCShape, IRoundedRectangle
+internal sealed class SCRoundedRectangle : SCAutoShape, IRoundedRectangle
 {
-    public SCRoundedRectangle(AutoShapeCollection autoShapeCollection, P.Shape pShape, SCGroupSCShape? groupShape) 
+    public SCRoundedRectangle(AutoShapeCollection autoShapeCollection, P.Shape pShape, SCGroupShape? groupShape) 
         : base(autoShapeCollection.ParentShapeCollection.ParentSlideObject, pShape, groupShape)
     {
     }
