@@ -8,11 +8,11 @@ namespace ShapeCrawler.Factories;
 
 internal sealed class AutoShapeCreator : OpenXmlElementHandler
 {
-    internal override SCShape? Create(OpenXmlCompositeElement pShapeTreeChild, OneOf<SCSlide, SCSlideLayout, SCSlideMaster> slideObject, SCGroupSCShape? groupShape)
+    internal override SCShape? Create(OpenXmlCompositeElement pShapeTreeChild, OneOf<SCSlide, SCSlideLayout, SCSlideMaster> slideObject, SCGroupShape? groupShape)
     {
         if (pShapeTreeChild is P.Shape pShape)
         {
-            var autoShape = new AutoSCShape(slideObject, pShape, groupShape);
+            var autoShape = new SCAutoShape(slideObject, pShape, groupShape);
             return autoShape;
         }
 
