@@ -20,7 +20,7 @@ internal static class FontDataParser
         }
 
         var placeholder = (SCPlaceholder)shape.Placeholder;
-        var autoShape = (AutoSCShape?)placeholder.ReferencedShape.Value;
+        var autoShape = (SCAutoShape?)placeholder.ReferencedShape.Value;
         autoShape?.FillFontData(paragraph.Level, ref phFontData);
     }
     
@@ -34,7 +34,7 @@ internal static class FontDataParser
             return fontData;
         }
         
-        var referencedShape = (AutoSCShape?)placeholder.ReferencedShape.Value;
+        var referencedShape = (SCAutoShape?)placeholder.ReferencedShape.Value;
         
         if (referencedShape is null && placeholder.Type is SCPlaceholderType.Title or SCPlaceholderType.CenteredTitle)
         {
